932 - Fix tests
